1) Deleted field injection Driver 2) Adding BasePage with common base page for pages that extends this class 3) deleted isTrue() from all methods where it was used
