+++ src/test/java/com/epam/tc/hw3/TestBase.java
@@ -1,11 +1,13 @@
 package com.epam.tc.hw3;
 
-import com.epam.tc.hw3.injectionDriver.WebDriverInjector;
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.assertj.core.api.SoftAssertions;
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
+
+import java.util.concurrent.TimeUnit;
 
 public class TestBase {
 
@@ -16,7 +18,8 @@ public class TestBase {
     public void setWebDriver() {
         WebDriverManager.chromedriver().setup();
         softAssertions = new SoftAssertions();
-        webDriver = new WebDriverInjector().getDriverService().getWebDriver();
+        webDriver = new ChromeDriver();
+        webDriver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
         webDriver.manage().window().maximize();
     }
