--- src/test/java/com/epam/tc/hw3/injectionDriver/DriverInjector.java
@@ -1,6 +1,0 @@
-package com.epam.tc.hw3.injectionDriver;
-
-public interface DriverInjector {
-
-    DriverService getDriverService();
-}
