--- src/test/java/com/epam/tc/hw3/injectionDriver/DriverService.java
@@ -1,7 +1,0 @@
-package com.epam.tc.hw3.injectionDriver;
-
-import org.openqa.selenium.WebDriver;
-
-public interface DriverService {
-    WebDriver getWebDriver();
-}
