--- src/test/java/com/epam/tc/hw3/injectionDriver/WebDriverInjector.java
@@ -1,8 +1,0 @@
-package com.epam.tc.hw3.injectionDriver;
-
-public class WebDriverInjector implements DriverInjector {
-    @Override
-    public DriverService getDriverService() {
-        return new DriverManager();
-    }
-}
